refactor(signals): Consolidate user signals
