+++ cl/api/models.py
@@ -1,19 +1,9 @@
 from http import HTTPStatus
 
-from django.conf import settings
 from django.contrib.auth.models import User
 from django.db import models
-from django.db.models.signals import post_save
-from django.dispatch import receiver
-from rest_framework.authtoken.models import Token
 
 from cl.lib.models import AbstractDateTimeModel
-
-
-@receiver(post_save, sender=settings.AUTH_USER_MODEL)
-def create_auth_token(sender, instance=None, created=False, **kwargs):
-    if created:
-        Token.objects.create(user=instance)
 
 
 class WebhookEventType(models.IntegerChoices):
